+++ src/main/java/io/fabric8/maven/docker/util/Logger.java
@@ -17,7 +17,6 @@ public interface Logger {
      */
     void debug(String format, Object ... params);
 
-
     /**
      * Informational message
      *