+++ doc/changelog.md
@@ -5,6 +5,7 @@
   - Break push with dedicated registry if temporary image tag already exists (#575)
   - Reduce log output for the non color case when pulling images (#568)
   - Add possibility to change colors in log messages
+  - Don't print a progressbar when in batch mode (mvn -B) (#564)
   
 * **0.16.5** (2016-09-27)
   - Refactored Docker connection parameter detection